projects/final/Dutt_Tiurina_Prusak/code/src/requirements.txt
pandas
numpy
# pydantic
# pydantic_ai 
tqdm
langchain_community
scikit-learn
torch
transformers
umap-learn
hdbscan